requirements/requirements_DEV.txt
aiohttp==3.8.1
aiomysql==0.1.1
aiosignal==1.2.0
alembic==1.8.1
alembic-autogen-check==1.1.1
amqp==5.1.1
anyio==3.6.1
argcomplete==2.0.0
async-timeout==4.0.2
attrs==22.1.0
azure-core==1.26.1
azure-storage-blob==12.14.1
billiard==3.6.4.0
black==22.10.0
celery==5.2.7
celery-redbeat==2.0.0
certifi==2022.9.24
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cryptography==38.0.4
datamodel-code-generator==0.11.19
Deprecated==1.2.13
dnspython==2.2.1
ecdsa==0.18.0
elasticsearch==7.17.6
elasticsearch-dsl==7.4.0
email-validator==1.3.0
fastapi==0.87.0
fastapi-code-generator==0.3.5
frozenlist==1.3.1
genson==1.2.2
greenlet==1.1.3
h11==0.13.0
httpcore==0.16.2
httpx==0.23.1
idna==3.3
inflect==5.6.2
isodate==0.6.1
isort==5.10.1
Jinja2==3.1.2
jsonschema==3.2.0
kombu==5.2.4
Mako==1.2.4
MarkupSafe==2.1.1
msrest==0.7.1
multidict==6.0.2
mypy==0.961
mypy-extensions==0.4.3
oauthlib==3.2.2
openapi-schema-validator==0.1.6
openapi-spec-validator==0.3.3
orjson==3.8.2
packaging==21.3
passlib==1.7.4
pathspec==0.10.2
platformdirs==2.5.4
prance==0.22.11.4.0
prompt-toolkit==3.0.31
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
PyMySQL==1.0.2
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.16.1
PySnooper==0.5.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.2.1
PyYAML==6.0
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
semver==2.13.0
shellingham==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a24
sse-starlette==1.1.6
starlette==0.21.0
stringcase==1.2.0
tenacity==8.1.0
toml==0.10.2
tomli==2.0.1
sqlacodegen==2.3.0
typed-ast==1.5.4
typer==0.4.2
types-PyMySQL==1.0.19
types-redis==4.3.4
typing_extensions==4.3.0
urllib3==1.26.13
uvicorn==0.18.3
vine==5.0.0
wcwidth==0.2.5
websocket-client==1.3.3
websockets==10.3
wrapt==1.14.1
XTTOOLS==1.3
yarl==1.8.1
types-python-dateutil
types-pytz
pypdf
python-barcode
reportlab==3.6.12
aiosmtplib
